remove extended examples
